Create README.md
Deploy a simple contract on EVM's L1 & L2 using Remix

Download this Workspace Satoshi Backup.zip

Goto Remix IDE by clicking here

Follow this procedure to upload the downloaded file image

Follow the process image

The next step here image

Unlock your Metamask wallet and Connect to Goereli Testnet image

Go to your metamask wallet and approve transaction. wait 30sec - 1 minutes to create contract Click the (view on etherscan) link in the console below the remix screen etherscan

You would see this image

Contract has been successfully deployed image

WARNING

This code is being provided as is. No guarantee, representation or warranty is being made, express or implied, as to the safety or correctness of the code. It has not been audited and as such there can be no assurance it will work as intended, and users may experience delays, failures, errors, omissions or loss of transmitted information. Nothing in this repo should be construed as investment advice or legal advice for any particular facts or circumstances and is not meant to replace competent counsel. It is strongly advised for you to contact a reputable attorney in your jurisdiction for any questions or concerns with respect thereto. a16z is not liable for any use of the foregoing, and users should proceed with caution and use at their own risk. See a16z.com/disclosures for more info.